Treasure Island
Treasure Island exercise.
Note to myself: Do not forget to use lower() function when asking something using input() if I need to compare the result. If the input is not lowered, "word" will return True, but "Word" or "WORD" will return false, which is not good for the user experience!