Password Generator
My solution of the password generator before looking at the solution from Angela.

I used a string.replace() function with a for loop to randomize the characters. Is it equivalent to Angela method in "randomness" ?

Angela used random.shuffle method which is simpler. I will redo it with Angela method